backbone_example/requirements.txt
Django==1.3.2
mimeparse>=0.1.3
python-dateutil==1.4
django-tastypie==0.9.11
pystache==0.3.1
pillow==3.0.0
requests==2.3.0
scikit-image
scikit-learn
scipy
